third_party/blink/web_tests/css3/flexbox/flexbox-ignore-container-firstLetter.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
    .container::first-letter { line-height: 100px; }
    .flexboxFirstLetter::first-letter { line-height: 200px; }
    p { line-height: 20px; }
</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.container')">
<div id=log></div>
<p>This test flex item should ignore container's firstLetter pseudo element.</p>

<div class="container">
    <div class="flexbox flexboxFirstLetter">
        <p data-expected-height=20>The first item.</p>
        <p data-expected-height=20>The second item.</p>
    </div>
    <p data-expected-height=20>Out of flex.</p>
</div>

<div class="container">
    <div class="flexbox">
        <p data-expected-height=20>The first item.</p>
        <p data-expected-height=20>The second item.</p>
    </div>
    <p data-expected-height=20>Out of flex.</p>
</div>
</body>
</html>
